server-app/requirements.txt
albumentations==1.4.15
dlib==19.24.6
facenet_pytorch==2.6.0
filterpy==1.4.5
Flask==3.0.3
insightface==0.7.3
lap==0.4.0
matplotlib==3.9.2
numpy==2.1.1
onnx==1.16.2
onnx_graphsurgeon==0.5.2
onnxsim==0.4.36
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
pafy==0.5.5
pandas==2.2.2
Pillow==10.4.0
psutil==6.0.0
pytorch_lightning==2.4.0
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
setuptools==70.0.0
skimage==0.0
tensorflow==2.17.0
tensorflow_intel==2.17.0
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
wandb==0.18.1
watchdog==5.0.2
